+ Refined worker queue time estimate.

+++ maestro/src/ibis/maestro/Service.java
@@ -16,8 +16,8 @@ import java.util.ArrayList;
  */
 public class Service
 {
-    static final int MICROSECOND_IN_NANOSECONDS = 1000;
-    static final int MILLISECOND_IN_NANOSECONDS = 1000*MICROSECOND_IN_NANOSECONDS;
+    static final long MICROSECOND_IN_NANOSECONDS = 1000L;
+    static final long MILLISECOND_IN_NANOSECONDS = 1000*MICROSECOND_IN_NANOSECONDS;
 
     /** Prints the platform version that is used. */
     static String getPlatformVersion()
